@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -16,7 +16,6 @@ dependencies:
 
   value_state: 1.5.1
   flutter_value_state: 1.3.6
-  value_cubit: 1.3.3
 dev_dependencies:
   flutter_test:
     sdk: flutter
