@@ -5,14 +5,15 @@ homepage: https://github.com/devobs
 version: 1.3.6
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   value_state: ^1.5.1
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
