@@ -5,7 +5,7 @@ homepage: https://github.com/devobs
 version: 1.3.3
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   bloc: ^8.1.0
