@@ -5,7 +5,7 @@ homepage: https://github.com/devobs
 version: 1.5.1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   meta: ^1.7.0
