@@ -1,24 +1,21 @@
-import 'package:value_cubit/value_cubit.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_value_state/flutter_value_state.dart';
 
-class CounterCubit extends ValueCubit<int> {
+class CounterCubit extends Cubit<Value<int>> {
   var _value = 0;
   Future<int> _getMyValueFromRepository() async => _value++;
 
-  CounterCubit() {
+  CounterCubit() : super(Value.success(0)) {
     increment();
   }
 
-  Future<void> increment() => perform(() async {
-        await Future.delayed(const Duration(seconds: 1));
-
+  Future<void> increment() => state.fetch(() async {
         final result = await _getMyValueFromRepository();
 
         if (result == 2) {
           throw 'Error';
-        } else if (result > 4) {
-          emit(const NoValueState());
         } else {
-          emit(ValueState(result));
+          return Value.success(result);
         }
-      });
+      }).forEach(emit);
 }
