@@ -1,53 +1,44 @@
 import 'package:flutter/widgets.dart';
 import 'package:value_state/value_state.dart';
 
-Widget _defaultBuilder<T>(BuildContext context, BaseState<T> state) =>
+Widget _defaultBuilder<T>(BuildContext context, Value<T> state) =>
     const SizedBox.shrink();
 
-Widget _defaultWrapper<T>(
-        BuildContext context, BaseState<T> state, Widget child) =>
+Widget _defaultWrapper<T>(BuildContext context, Value<T> state, Widget child) =>
     child;
 
 typedef OnValueStateWaiting<T> = Widget Function(
-    BuildContext context, WaitingState<T> state);
+    BuildContext context, Value<T> state);
 
 typedef OnValueStateWithValue<T> = Widget Function(
-    BuildContext context, WithValueState<T> state, Widget? error);
-typedef OnValueStateNoValue<T> = Widget Function(
-    BuildContext context, NoValueState<T> state);
+    BuildContext context, Value<T> state, Widget? error);
 typedef OnValueStateError<T> = Widget Function(
-    BuildContext context, ErrorState<T> state);
+    BuildContext context, Value<T> state);
 typedef OnValueStateDefault<T> = Widget Function(
-    BuildContext context, BaseState<T> state);
+    BuildContext context, Value<T> state);
 typedef OnValueStateWrapper<T> = Widget Function(
-    BuildContext context, BaseState<T> state, Widget child);
+    BuildContext context, Value<T> state, Widget child);
 
-/// Define default behavior for the states [WaitingState], [NoValueState], [ErrorState].
+/// Define default behavior for the states init, success and error.
 /// [builderDefault] can be used when none of this callback is mentionned.
-class ValueStateConfigurationData {
-  const ValueStateConfigurationData({
+class ValueBuilderConfigurationData {
+  const ValueBuilderConfigurationData({
     OnValueStateWrapper? wrapper,
-    OnValueStateWaiting? builderWaiting,
-    OnValueStateNoValue? builderNoValue,
-    OnValueStateError? builderError,
+    OnValueStateWaiting? initial,
+    OnValueStateError? failure,
     OnValueStateDefault? builderDefault,
   })  : _wrapper = wrapper,
-        _builderWaiting = builderWaiting,
-        _builderNoValue = builderNoValue,
-        _builderError = builderError,
+        _builderWaiting = initial,
+        _builderError = failure,
         _builderDefault = builderDefault;
 
   /// Builder for all states that will be wrapped by this builder.
   OnValueStateWrapper get wrapper => _wrapper ?? _defaultWrapper;
   final OnValueStateWrapper? _wrapper;
 
-  /// Builder for [WaitingState].
+  /// Builder for [Value].
   OnValueStateWaiting get builderWaiting => _builderWaiting ?? builderDefault;
   final OnValueStateWaiting? _builderWaiting;
-
-  /// Builder for [NoValueState].
-  OnValueStateNoValue get builderNoValue => _builderNoValue ?? builderDefault;
-  final OnValueStateNoValue? _builderNoValue;
 
   /// Builder for [ErrorState].
   OnValueStateError get builderError => _builderError ?? builderDefault;
@@ -57,34 +48,31 @@ class ValueStateConfigurationData {
   OnValueStateDefault get builderDefault => _builderDefault ?? _defaultBuilder;
   final OnValueStateDefault? _builderDefault;
 
-  /// Creates a copy of this [ValueStateConfigurationData] but with the given
+  /// Creates a copy of this [ValueBuilderConfigurationData] but with the given
   /// fields replaced with the new values.
-  ValueStateConfigurationData copyWith({
+  ValueBuilderConfigurationData copyWith({
     OnValueStateWrapper? wrapper,
     OnValueStateWaiting? builderWaiting,
-    OnValueStateNoValue? builderNoValue,
     OnValueStateError? builderError,
     OnValueStateDefault? builderDefault,
   }) =>
-      ValueStateConfigurationData(
+      ValueBuilderConfigurationData(
         wrapper: wrapper ?? this.wrapper,
-        builderWaiting: builderWaiting ?? this.builderWaiting,
-        builderNoValue: builderNoValue ?? this.builderNoValue,
-        builderError: builderError ?? this.builderError,
+        initial: builderWaiting ?? this.builderWaiting,
+        failure: builderError ?? this.builderError,
         builderDefault: builderDefault ?? this.builderDefault,
       );
 
-  /// Creates a new [ValueStateConfigurationData] where each parameter
+  /// Creates a new [ValueBuilderConfigurationData] where each parameter
   /// from this object has been merged with the matching attribute.
-  ValueStateConfigurationData merge(
-      ValueStateConfigurationData? configuration) {
+  ValueBuilderConfigurationData merge(
+      ValueBuilderConfigurationData? configuration) {
     final baseConfiguration =
-        configuration ?? const ValueStateConfigurationData();
+        configuration ?? const ValueBuilderConfigurationData();
 
     return baseConfiguration.copyWith(
       wrapper: _wrapper,
       builderWaiting: _builderWaiting,
-      builderNoValue: _builderNoValue,
       builderError: _builderError,
       builderDefault: _builderDefault,
     );
@@ -94,37 +82,35 @@ class ValueStateConfigurationData {
   bool operator ==(other) =>
       identical(this, other) ||
       runtimeType == other.runtimeType &&
-          other is ValueStateConfigurationData &&
+          other is ValueBuilderConfigurationData &&
           wrapper == other.wrapper &&
           builderWaiting == other.builderWaiting &&
-          builderNoValue == other.builderNoValue &&
           builderError == other.builderError &&
           builderDefault == other.builderDefault;
 
   @override
   int get hashCode => Object.hash(
         wrapper,
-        builderNoValue,
         builderWaiting,
         builderError,
         builderDefault,
       );
 }
 
-/// Provide a [ValueStateConfigurationData] for all inherited widget to define
-/// default behavior for any state of [BaseState] except [ValueState].
+/// Provide a [ValueBuilderConfigurationData] for all inherited widget to define
+/// default behavior for any state of [Value] except [ValueState].
 ///
-/// If this configuration is in a subtree of another [ValueStateConfiguration],
+/// If this configuration is in a subtree of another [ValueBuilderConfiguration],
 /// the configuration will be merged with the parent one.
-class ValueStateConfiguration extends StatelessWidget {
-  const ValueStateConfiguration({
+class ValueBuilderConfiguration extends StatelessWidget {
+  const ValueBuilderConfiguration({
     super.key,
     required this.configuration,
     required this.child,
   });
 
   /// The default to configuration.
-  final ValueStateConfigurationData configuration;
+  final ValueBuilderConfigurationData configuration;
   final Widget child;
 
   @override
@@ -136,15 +122,15 @@ class ValueStateConfiguration extends StatelessWidget {
         child: child);
   }
 
-  static ValueStateConfigurationData? maybeOf(BuildContext context) => context
+  static ValueBuilderConfigurationData? maybeOf(BuildContext context) => context
       .dependOnInheritedWidgetOfExactType<_ValueStateConfiguration>()
       ?.configuration;
 
-  static ValueStateConfigurationData of(BuildContext context) {
-    final ValueStateConfigurationData? configuration = maybeOf(context);
+  static ValueBuilderConfigurationData of(BuildContext context) {
+    final ValueBuilderConfigurationData? configuration = maybeOf(context);
 
-    assert(
-        configuration != null, 'No $ValueStateConfiguration found in context');
+    assert(configuration != null,
+        'No $ValueBuilderConfiguration found in context');
 
     return configuration!;
   }
@@ -154,7 +140,7 @@ class _ValueStateConfiguration extends InheritedWidget {
   const _ValueStateConfiguration(
       {required this.configuration, required super.child});
 
-  final ValueStateConfigurationData configuration;
+  final ValueBuilderConfigurationData configuration;
 
   @override
   bool updateShouldNotify(covariant _ValueStateConfiguration oldWidget) =>
