@@ -2,30 +2,32 @@ import 'package:test/test.dart';
 import 'package:value_state/value_state.dart';
 
 void main() {
-  test('perform on $ValueState', () {
-    final stream =
-        const ValueState<int>(1).perform((state) async => const ValueState(2));
+  test('perform on ${Value<int>}', () {
+    final stream = Value.success(1).fetch(() async => Value.success(2));
 
     expect(
         stream,
         emitsInOrder([
-          const ValueState(1, refreshing: true),
-          const ValueState(2, refreshing: false),
+          Value.success(1, isFetching: true),
+          Value.success(2, isFetching: false),
         ]));
   });
 
-  test('performStream on $ValueState', () {
-    final stream = const ValueState<int>(1).performStream((state) async* {
-      yield const ValueState(2);
-      yield const ValueState(3);
-    });
+  test('performStream on ${Value<int>}', () {
+    final stream = Value.success(1).fetchStream(
+      Stream.fromIterable([
+        Value.success(2),
+        Value.success(3),
+      ]),
+    );
 
     expect(
-        stream,
-        emitsInOrder([
-          const ValueState(1, refreshing: true),
-          const ValueState(2, refreshing: false),
-          const ValueState(3, refreshing: false),
-        ]));
+      stream,
+      emitsInOrder([
+        Value.success(1, isFetching: true),
+        Value.success(2, isFetching: false),
+        Value.success(3, isFetching: false),
+      ]),
+    );
   });
 }
